Move classes to other packages Resolve Sonar issues Update documentation
