+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/RepositoryValidationDecorator.java
@@ -6,6 +6,7 @@ import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.support.QueryImpl;
 import org.molgenis.data.validation.data.AttributeValueConstraint;
 import org.molgenis.data.validation.data.AttributeValueValidationResult;
+import org.molgenis.data.validation.data.EntityAttributesValidator;
 import org.molgenis.util.HugeMap;
 import org.molgenis.util.HugeSet;
