+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/ValidationMessage.java
@@ -2,9 +2,13 @@ package org.molgenis.data.validation;
 
 import com.google.auto.value.AutoValue;
 
+import java.io.Serializable;
+
 @AutoValue
-public abstract class ValidationMessage
+public abstract class ValidationMessage implements Serializable
 {
+	private static final long serialVersionUID = 1L;
+
 	public abstract String getErrorCode();
 
 	public abstract String getMessage();
