+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/ValidationResult.java
@@ -1,5 +1,11 @@
 package org.molgenis.data.validation;
 
+/**
+ * @see org.molgenis.data.validation.meta.AttributeValidationResult
+ * @see org.molgenis.data.validation.meta.EntityTypeValidationResult
+ * @see org.molgenis.data.validation.meta.PackageValidationResult
+ * @see CompositeValidationResult
+ */
 public interface ValidationResult
 {
 	boolean hasConstraintViolations();
